Update bonspiels.html
We started strong, scoring 2 to start the game. Even though we gave up a big second, it was still a 1-point game at the half. We weren't able to string together enough rocks to stay in it though, giving up another big end and not leveraging any big ends for ourselves.