Frogtown Game 3 Results
We struggled with draw weight a bit to start the game as well as the shape of the sheet, giving up a 2 then getting forced to start the game. We stole 2 to get right back in it but then couldn't quite get takeouts to do what we wanted to and gave back 3. After a back-and-forth end where we couldn't get their rock off the back of the button, we had to get 3 in the last to force a shootout tiebreaker but our last shot never curled off a guard and we only scored 2.